pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "moogla"
version = "0.0.1"
description = "Local-first LLM runtime and orchestration framework"
authors = [{name="Rehkad"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9",
    "fastapi>=0.110",
    "uvicorn>=0.28",
    "openai>=1.30",
    "httpx>=0.27,<0.28",
    "sqlmodel>=0.0.16",
    "passlib[bcrypt]>=1.7.4",
    "python-jose>=3.3",
]

[project.scripts]
moogla = "moogla.cli:app"

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit",
    "httpx>=0.27,<0.28"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"moogla.web" = ["*"]

